test: ensure service is up before running Check, ListObjects, and WriteModel tests
