+++ pkg/server/server.go
@@ -931,8 +931,8 @@ func (s *Server) ListStores(ctx context.Context, req *openfgav1.ListStoresReques
 	return q.Execute(ctx, req)
 }
 
-// IsReady reports whether this OpenFGA server instance is ready to accept
-// traffic.
+// IsReady reports whether the datastore is ready. Please see the implementation of [[storage.OpenFGADatastore.IsReady]]
+// for your datastore.
 func (s *Server) IsReady(ctx context.Context) (bool, error) {
 	// for now we only depend on the datastore being ready, but in the future
 	// server readiness may also depend on other criteria in addition to the
