+++ pkg/testutils/testutils.go
@@ -2,14 +2,25 @@
 package testutils
 
 import (
+	"context"
+	"errors"
+	"fmt"
 	"math/rand"
 	"sort"
 	"strconv"
 	"testing"
+	"time"
 
+	"github.com/cenkalti/backoff/v4"
 	"github.com/google/go-cmp/cmp"
+	"github.com/hashicorp/go-retryablehttp"
 	openfgav1 "github.com/openfga/api/proto/openfga/v1"
 	"github.com/stretchr/testify/require"
+	"google.golang.org/grpc"
+	grpcbackoff "google.golang.org/grpc/backoff"
+	"google.golang.org/grpc/credentials"
+	"google.golang.org/grpc/credentials/insecure"
+	healthv1pb "google.golang.org/grpc/health/grpc_health_v1"
 	"google.golang.org/protobuf/types/known/structpb"
 )
 
@@ -80,4 +91,59 @@ func MakeStringWithRuneset(n uint64, runeSet []rune) string {
 	}
 
 	return s
+}
+
+// EnsureServiceUp is a test helper that ensures that a service's grpc health endpoint is responding OK. It can also
+// ensure that the HTTP /healthz endpoint is responding OK.
+func EnsureServiceUp(t testing.TB, grpcAddr, httpAddr string, transportCredentials credentials.TransportCredentials, httpHealthCheck bool) {
+	t.Helper()
+
+	timeoutCtx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
+	defer cancel()
+
+	creds := insecure.NewCredentials()
+	if transportCredentials != nil {
+		creds = transportCredentials
+	}
+
+	dialOpts := []grpc.DialOption{
+		grpc.WithBlock(),
+		grpc.WithTransportCredentials(creds),
+		grpc.WithConnectParams(grpc.ConnectParams{Backoff: grpcbackoff.DefaultConfig}),
+	}
+
+	conn, err := grpc.DialContext(
+		timeoutCtx,
+		grpcAddr,
+		dialOpts...,
+	)
+	require.NoError(t, err)
+	defer conn.Close()
+
+	client := healthv1pb.NewHealthClient(conn)
+
+	policy := backoff.NewExponentialBackOff()
+	policy.MaxElapsedTime = 10 * time.Second
+
+	err = backoff.Retry(func() error {
+		resp, err := client.Check(timeoutCtx, &healthv1pb.HealthCheckRequest{
+			Service: openfgav1.OpenFGAService_ServiceDesc.ServiceName,
+		})
+		if err != nil {
+			return err
+		}
+
+		if resp.GetStatus() != healthv1pb.HealthCheckResponse_SERVING {
+			return errors.New("not serving")
+		}
+
+		return nil
+	}, policy)
+	require.NoError(t, err)
+
+	if httpHealthCheck {
+		resp, err := retryablehttp.Get(fmt.Sprintf("http://%s/healthz", httpAddr))
+		require.Equal(t, 200, resp.StatusCode)
+		require.NoError(t, err)
+	}
 }
